update pipeline scripts to call Invoke-ImageBuilder

+++ scripts/Invoke-ImageBuilder.ps1
@@ -7,6 +7,9 @@ Executes ImageBuilder with the specified args.
 .PARAMETER ImageBuilderArgs
 The args to pass to ImageBuilder.
 
+.PARAMETER DockerRunOptions
+Additional options to pass to the "docker run" command.
+
 .PARAMETER ReuseImageBuilderImage
 Indicates that a previously built ImageBuilder image is presumed to exist locally and that 
 it should be used for this execution of the script.  This allows some optimization when
@@ -25,6 +28,9 @@ The ScriptBlock is passed the following argument values:
 param(
     [string]
     $ImageBuilderArgs,
+
+    [string]
+    $DockerRunOptions,
 
     [switch]
     $ReuseImageBuilderImage,
@@ -59,6 +65,9 @@ try {
 
     $imageNames = ./Get-ImageNames.ps1
 
+    # If we need to execute the OnCommandExcecuted callback, defer the removal of the container until the end of the script.
+    $deferContainerRemoval = ($null -ne $OnCommandExecuted)
+
     $activeOS = docker version -f "{{ .Server.Os }}"
     if ($activeOS -eq "linux") {
         # On Linux, ImageBuilder is run within a container.
@@ -67,8 +76,14 @@ try {
             ./scripts/Invoke-WithRetry "docker pull $($imageNames.ImageBuilderLinux)"
             Exec "docker build -t $imageBuilderImageName --build-arg IMAGE=$($imageNames.ImageBuilderLinux) -f ./scripts/Dockerfile.WithRepo ."
         }
-        
-        $imageBuilderCmd = "docker run --name $imageBuilderContainerName -v /var/run/docker.sock:/var/run/docker.sock $imageBuilderImageName"
+
+        $dockerRunOptions = "$DockerRunOptions --name $imageBuilderContainerName -v /var/run/docker.sock:/var/run/docker.sock"
+
+        if (-not $deferContainerRemoval) {
+            $dockerRunOptions += " --rm"
+        }
+
+        $imageBuilderCmd = "docker run $dockerRunOptions $imageBuilderImageName"
     }
     else {
         # On Windows, ImageBuilder is run locally due to limitations with running Docker client within a container.
@@ -78,8 +93,7 @@ try {
             ./scripts/Invoke-WithRetry "docker pull $($imageNames.ImageBuilderWindows)"
             Exec "docker create --name $imageBuilderContainerName $($imageNames.ImageBuilderWindows)"
 
-            if (Test-Path -Path $imageBuilderFolder)
-            {
+            if (Test-Path -Path $imageBuilderFolder) {
                 Remove-Item -Recurse -Force -Path $imageBuilderFolder
             }
 
@@ -94,7 +108,9 @@ try {
     }
 }
 finally {
-    Exec "docker container rm -f $imageBuilderContainerName"
+    if ($deferContainerRemoval) {
+        Exec "docker container rm -f $imageBuilderContainerName"
+    }
     
     popd
 }
